# MiqExpression#to_ruby Date bug fix and speed improvements