### Fix Bug

The `;` in the expression was screwing up the logic

```
col > 1 and col > 2 and val=col ; !val.nil? && val > 1 && val < 2
# =>
val > 1 && val < 2
```

It ignores the part before the semi so it skips most of the miq expression

Solution is to not use a semi

```
col > 1 and col > 2 and !(val=col).nil? and val > 1 and val < 2
```

Solution: Got rid of the semi. used a consistent operator

### Speedup date creation

Serialization from String#to_time just took too much time.

When running MiqExpression in ruby, the `to_time` takes up 411 of 530ms.
Using a more ruby friendly form of dates in ruby speeds this up.

| method      | before | after | change |
|-------------|--------|-------|--------|
|         all |   2796 |  2380 |    15% |
|     do_eval |    530 |   172 |    68% |
|     to_time |    411 |    15 |    96% |

Sorry, for benchmark, only ran test twice for before/after (throwing away the first run) The timings in `do_eval` seemed stable in the testing and didn't need too much riggor.

For the time case:

```
mexp = MiqExpression.new(
  "AND" => [
    {">="  => {"field" => "Vm-id", "value" => "1"}},
    {">="  => {"field" => "Vm-host_id", "value" => "1"}},
    {"="   => {"field" => "Vm-vendor_display", "value" => "VMware"}},
    {"="   => {"field" => "Vm-vendor", "value" => "vmware"}},
    {">="  => {"field" => "Vm-created_on", "value" => 5.years.ago}},
    {"<"   => {"field" => "Vm-created_on", "value" => 5.years.from_now}},
  ])
```

`to_ruby` (in the after case) yields:

```
(<value ref=vm, type=integer>/virtual/id</value> >= 1 and
 <value ref=vm, type=integer>/virtual/host_id</value> >= 1 and
 <value ref=vm, type=string>/virtual/vendor_display</value> == "VMware" and
 <value ref=vm, type=string>/virtual/vendor</value> == "vmware" and
 <value ref=vm, type=datetime>/virtual/created_on</value> >= Time.utc(2018, 2, 14, 15, 35, 29) and
 <value ref=vm, type=datetime>/virtual/created_on</value> < Time.utc(2028, 2, 14, 15, 35, 29)
)
```

Most of the performance improvements we were making were about reducing number of fields in the expression. This is the first one to speed up the performance of the expression it self.

### No Date fields

We do not have any full tests around `:date` datatype because we don't have any. All are DateTime.

Also, for some reason, we are showing some time spent in coercing DateTime and time. which is confusing since we are only dealing with Time objects.