@kbrock 2023-02-15
update:
- removed `Date.to_date`.

There are no date columns in our codebase, so we don't have any tests around :date.
I'm thinking this is fine for now.