@kbrock 2023-02-21
udpate:
- removed extra space to put comment near method definition
- fixed comment description to properly state the output of the method